Remove duplicate DocumentFileUtils
