+++ gateway-export-plugin/src/main/java/com/ca/apim/gateway/cagatewayexport/util/injection/ExportPluginModule.java
@@ -6,6 +6,7 @@
 
 package com.ca.apim.gateway.cagatewayexport.util.injection;
 
+import com.ca.apim.gateway.cagatewayconfig.util.file.DocumentFileUtils;
 import com.ca.apim.gateway.cagatewayexport.tasks.explode.ExplodeBundle;
 import com.ca.apim.gateway.cagatewayexport.tasks.explode.bundle.entity.EntityTypeRegistry;
 import com.ca.apim.gateway.cagatewayexport.tasks.explode.filter.EntityFilter;
@@ -16,7 +17,6 @@ import com.ca.apim.gateway.cagatewayexport.tasks.explode.loader.EntityLoader;
 import com.ca.apim.gateway.cagatewayexport.tasks.explode.loader.EntityLoaderRegistry;
 import com.ca.apim.gateway.cagatewayexport.tasks.explode.writer.EntityWriter;
 import com.ca.apim.gateway.cagatewayexport.tasks.explode.writer.EntityWriterRegistry;
-import com.ca.apim.gateway.cagatewayexport.util.file.DocumentFileUtils;
 import com.ca.apim.gateway.cagatewayexport.util.json.JsonTools;
 import com.ca.apim.gateway.cagatewayexport.util.policy.PolicyXMLSimplifier;
 import com.ca.apim.gateway.cagatewayexport.util.xml.DocumentTools;
